backend/reports/serializers.py
from rest_framework import serializers

from .models import ReportSnapshot


class ReportSnapshotSerializer(serializers.ModelSerializer):
    class Meta:
        model = ReportSnapshot
        fields = ["id", "name", "created_at", "data"]

