backend/requirements.txt
django>=5.1
djangorestframework>=3.16
django-cors-headers>=4.9
django-environ>=0.13
